+++ lib/grape/validations/validators/exactly_one_of_validator.rb
@@ -11,11 +11,10 @@ module Grape
         end
 
         def validate_params!(params)
-          known_keys = all_keys
-          keys = keys_in_common(params, known_keys)
+          keys = keys_in_common(params)
           return if keys.length == 1
 
-          validation_error!(known_keys, @exactly_one_exception_message) if keys.empty?
+          validation_error!(all_keys, @exactly_one_exception_message) if keys.empty?
           validation_error!(keys, @mutual_exclusion_exception_message)
         end
       end